pyproject.toml
[tool.poetry]
name = "docs-compiler"
version = "0.1.0"
description = ""
authors = ["razzle <harry@razzle.cloud>"]

packages = [
    { include = "docs-compiler" }
]

[tool.poetry.scripts]
bb-docs-compiler = 'docs-compiler.cli:compiler'
bb-docs-info = 'docs-compiler.info:info'

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
GitPython = "^3.1.27"
Jinja2 = "^3.1.2"
"ruamel.yaml" = "^0.17.21"
pathlib = "^1.0.1"
python-frontmatter = "^1.0.0"
mkdocs-material = "^8.3.9"
mike = "^1.1.2"
mkdocs-awesome-pages-plugin = "^2.8.0"
deepmerge = "^1.0.1"
rich = "^12.5.1"
requests = "^2.28.1"
semver = "^2.13.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}
isort = {version = "^5.10.1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
